requirements.txt
numpy
pygame
pillow
svg.path